requirements.txt
# Core NLP Libraries
spacy>=3.5.0
nltk>=3.8.0

# Machine Learning & Vectorization
scikit-learn>=1.2.0
numpy>=1.24.0
pandas>=2.0.0

# Semantic Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=7.3.0
